Replace master with 'old-state'
'old-state' was originally branched as a reversion of broken commits in
order to get back to a working state. Development continued on that
branch rather than master or expressified.

I guess I had made small changes to the master branch locally and I
don't care to deal with merging the PR manually so this is just
replacing the master branch entirely with the base from 'old-state.'

TL;DR - I suck at git.