Fully injected BTTV
NOTE: BTTV is **NOT** included. It will have to be provided by the end
user. In later non-alpha (what's before alpha?) releases, package.json
will not contain the "inject_js_end" property. The user will have to
manually add this and supply the *.js file in order to get BTTV working.