Changes & implement tmi.js
- Reworked most of the console logging to use logger.js
- Removed node-twitch-api (it was basically request anyway) and replaced
it with tmi.js. I need tmi anyway for hosts & subs and it also does API
requests.
- Renamed the user info stored in the config file
- Added the user's Twitch _id to the auth callback and config storage
- Wrapped any twitch.js calls (just pollFollowers now but will contain
donations etc.) in an initAPI() function - which is what's called from
the server at startup.
- Removed the stupid type checking from resolveUser() since it wasn't
necessary and reduces the number of arguments it needs
- Increased time between follower alerts from 1.5 seconds to 2.5
- Cleaned up package.json

TODO: event storing + list, front-end alerts for hosts & subs, improve
the dashboard