{fix} Tock fixes
I discovered that Tock didn't actually function quite the way I thought
it did. I assumed it was clearing timeouts / intervals the way I
expected but since the function being called was a new instance as far
as it was concerned, it was never able to clear before creating a new
one. Multiple timeouts / intervals of the same function were firing as a
result.

I fixed this by changing the syntax to require a unique ID string as the
first argument in order to identify the function being called.