{refactor} bot -> core & events
The bot core is now a class that extends EventEmitter, meaning $.emit()
and $.on() are available ( along with the rest of EventEmitter's methods
).

This is now used for the command registration instead of Transit which
is now going to be used bot-side for things that communicate with the
main app, like `Transit.on('alert:follow')`.

As a result of these changes, things were restructured pretty heavily in
the core. The core modules needed to be loaded just after the core was
initiated but ES2015 imports are hoisted and can't be within blocks.
These core 'components' as I'm now calling them are now require'd
instead, in a _loadComponents() function when initializing the bot.

The 'core' folder containing these modules has been renamed to
'components.'