{refactor} Make db set & confirm return their values
set() and confirm() methods for the db ( and the bot core ) will now return values which can simplify code in some cases. 

This means they no longer return their context so cannot be chained but whatevs.

Also of note, they actually return the value from the db, so if for some reason the put / update fails, the old value will be returned.