{fix} db#getRecentFollows order logic
Order componenent needs to be passed as a property of the options object, not as an array.

Also commented out dbGetEvents and clarified the TODO. It's a super old, untouched function that never really worked in the first place, uses raw queries, and is generally crap. I'll get to it eventually.