feat(bot): add weave for bot response customization
Add `weave` which is available to all bot modules through the `$` API for string localization or customization.

Add a new `language` element to bot components that registers response strings.

This commit forms the backend for this system. There are a few more parts to flesh out, like exposing some kind of interface for the user to change strings, language overrides, etc.