feat: replace EventEmitter with EventEmitter2
Update transit and the bot core to extend EventEmitter2.  This allows for easily forwarding transit events through the bot core with the `onAny` method.

EE2 is more performant and has more features, like wildcarding events and a few extra methods.

Override the EE2 `on` method on the bot core to prevent duplicating event listeners. This was previously done as a non-standard in the events module but it's important for performance across the board. Since bot modules get their caches cleared and are reloaded when running their commands, event listeners could potentially be added and re-added and pile up. This will prevent that kind of memory leak throughout anything that uses the `$.on` method.