refactor(bot): flatten twitch-api component
Simplify & restructure. Use the `baseURL` feature of axios to default to Kraken endpoints if the provided URL is not absolute. Headers by default include token & client ID but can be overridden by providing `opts.headers` as an object in the second parameter to `api`. Replace a couple instances of `moment().valueOf()` with `Date.now()` since they return the same result.