fix(weave/bot): points lang path ambiguity
Setting a lang string to a path as a string like 'top' then setting another string to a path like 'top.inner' would overwrite 'top' due to it not being an object. This was causing a points response to fail and has been fixed.

Add handling for this case to weave which activates if the returned value is not a string.