refactor: vue-resource -> axios, formatting, ...
Replace vue-resource with axios in the vuex user module. This gets rid of errors in the renderer because vue-resource didn't handle URL params well ( or I did it wrong, but axios is used everywhere else anyway ).

Also added a few helpful comments to the bot twitch-api module and the helpers module.