feat: add tray, rewrite window manager, ...
Add an icon for the app in the system tray. Allows for hiding / showing the app. Add an app icon to windows.

Rewrite the window manager entirely. It had some useless & uselessly complex features, and is no longer a class but just a collection of functions. This was a hold-over from a boilerplate.

Introduce the `cubbie` module for managing state in the main process - works similarly to `vuex` on the renderer.

Add methods in transit to forward & send events to the renderer and to the bot.

Further normalize services to make them all work the same way.

Add settings configuration for the bot to the UI.

Implement a notification snackbar on the renderer for UI feedback. Add an event bus as a small Vue plugin.

Add a splash window on startup.