fix: twitch login, debug context menu, babel
Fix Twitch login - I'm not sure when it even broke since I've been logged in for a long time but it works now.

Remove debug context menu from production builds.

Change fast-async babel setting to use directives for adding the nodent runtime. This requiers a workaround on the renderer since fast-async won't actually inject the runtime if it doesn't need it, so there's an empty async function in the entry point.